+++ test/iterator.js
@@ -643,6 +643,45 @@ metatests.testSync('Iterator.chainApply on string', test => {
   test.strictSame(actual, 'h e l l o');
 });
 
+metatests.testSync('Iterator.keepLast empty', test => {
+  const actual = iter([]).keepLast();
+  test.strictSame(actual.last, undefined);
+  actual.repeatLast();
+  test.strictSame(actual.next(), { done: true, value: undefined });
+  test.strictSame(actual.next(), { done: true, value: undefined });
+});
+
+metatests.testSync('Iterator.keepLast repeat', test => {
+  const actual = iter([1, 2, 3, 4]).keepLast();
+  test.strictSame(actual.next(), { done: false, value: 1 });
+  test.strictSame(actual.last, 1);
+
+  actual.repeatLast();
+  test.strictSame(actual.last, 1);
+  test.strictSame(actual.next(), { done: false, value: 1 });
+  test.strictSame(actual.last, 1);
+
+  test.strictSame(actual.next(), { done: false, value: 2 });
+  test.strictSame(actual.next(), { done: false, value: 3 });
+
+  actual.repeatLast();
+  test.strictSame(actual.last, 3);
+  test.strictSame(actual.next(), { done: false, value: 3 });
+  test.strictSame(actual.last, 3);
+
+  test.strictSame(actual.next(), { done: false, value: 4 });
+  test.strictSame(actual.next(), { done: true, value: undefined });
+  test.strictSame(actual.last, undefined);
+});
+
+metatests.testSync('Iterator.keepLast takeWhile', test => {
+  const keeper = iter([1, 2, 3, 4]).keepLast();
+  test.strictSame(keeper.takeWhile(v => v < 3).toArray(), [1, 2]);
+  test.strictSame(keeper.last, 3);
+  keeper.repeatLast();
+  test.strictSame(keeper.toArray(), [3, 4]);
+});
+
 metatests.testSync('iterEntries must iterate over object entries', test => {
   const source = { a: 13, b: 42, c: 'hello' };
   test.strictSame(iterEntries(source).toArray(), Object.entries(source));